account adjust
Former-commit-id: aae6472ff80b527375f7dbe8ee7c3b7e70d03d34
Former-commit-id: 778ed878efb7a27c796ca3861635a1dbc5c9fbf1

+++ thrift/member_service.thrift
@@ -78,15 +78,15 @@ service MemberService{
       4:i32 amount,5:string outerNo,6:string remark)
     // 账户消耗,amount精确到分
     ttype.Result AccountConsume(1:i64 memberId,2:i32 account,3:string title,
-      4:double amount, 5:string outerNo,6:string remark)
+      4:i32 amount, 5:string outerNo,6:string remark)
     // 账户抵扣,amount精确到分
     ttype.Result AccountDiscount(1:i64 memberId,2:i32 account,3:string title,
-      4:double amount, 5:string outerNo,6:string remark)
+      4:i32 amount, 5:string outerNo,6:string remark)
     // 账户退款,amount精确到分
     ttype.Result AccountRefund(1:i64 memberId,2:i32 account,3:string title,
-        4:double amount, 5:string outerNo,6:string remark)
+        4:i32 amount, 5:string outerNo,6:string remark)
     // 账户人工调整
-    ttype.Result AccountAdjust(1:i64 memberId,2:i32 account,3:double value,4:i64 relateUser,5:string remark)
+    ttype.Result AccountAdjust(1:i64 memberId,2:i32 account,3:i32 value,4:i64 relateUser,5:string remark)
 
     // !银行四要素认证
     ttype.Result B4EAuth(1:i64 memberId,2:string action,3:map<string,string> data)
